Add files via upload
This is the project done be me and my co-coder .
about the project, we choose to make an investment calculator. In this calculator we have used the equation  S = P * (1+ (R/100)*T) to calculate the final amount of money which is invested, and we also took input like principle amount of deposit, year of investment, interest of money.  And finally the amount is calculated
We had used Visual Studio code to learn and do python project. 

To run this project -
    It is easy to use this project as our normal calculator. we just need to give input of our amount of deposit, interest, and year. then we need to click enter ,the calculator will calculate the final amount